% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi2.R
\name{model_based_pi0}
\alias{model_based_pi0}
\title{MAP fit of pi0 assuming non-null come from a beta distribution.}
\usage{
model_based_pi0(p, breaks = seq(0, 1, by = 0.01), alpha_prior = c(2,
  2), beta_prior = c(2, 1), ...)
}
\arguments{
\item{p}{Samples (p-values for pisquared's application)}

\item{alpha_prior}{Shape1 and shape2 of beta() prior on alpha. Shape1,shape2 > 1 prevents alpha collapsing to 0 or 1.}

\item{beta_prior}{Shape and rate of prior on beta_minus_one. Shape>1 enforces beta > 1.}

\item{...}{passed to rstan fitting function choosen by `method`}
}
\value{
Fitted model parameters. .
}
\description{
MAP fit of pi0 assuming non-null come from a beta distribution.
}
