% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi2.R
\name{pi0est_copy}
\alias{pi0est_copy}
\title{Copy of qvalue's pi0est default method. Included here to avoid dependency on bioconductor.}
\usage{
pi0est_copy(p, lambda = seq(0.05, 0.95, 0.05), smooth.df = 3)
}
\arguments{
\item{lambda}{The value of the tuning parameter to estimate pi_0. Must be in [0,1). Optional, see Storey (2002).}

\item{smooth.df}{Degrees-of-freedom to use when estimating pi_0 with a smoother. Optional.}
}
\description{
Assumes pi0.method="smoother" and smooth.log.pi0 = F (see qvalue docs).
}
