% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi2.R
\name{pi2_estimator}
\alias{pi2_estimator}
\title{Estimate sharing between two lists of p-values.}
\usage{
pi2_estimator(p1, p2, breaks = seq(0, 1, by = 0.01), restarts = 5)
}
\arguments{
\item{p1}{First list of p-values}

\item{p2}{Second list of p-values}

\item{breaks}{Grid for estimating p-value histogram}

\item{restarts}{Number of random restarts to perform.}

\item{alpha_prior}{Shape1 and shape2 of beta() prior on alpha. Shape1,shape2 > 1 prevents alpha collapsing to 0 or 1.}

\item{beta_prior}{Shape and rate of prior on beta_minus_one. Shape>1 enforces beta > 1.}

\item{...}{passed to rstan fitting function choosen by `method`}
}
\value{
List of jaccard index and fitted model parameters.
}
\description{
Note the result is symmetric in p1 and p2. This will not work well if for either list all the p-values are uniform
(i.e. from the null) or all the p-values are significant. Also note this method (like Storey's pi0) assumes
}
\details{
To run restarts in parallel registerDoMC needs to have been run beforehand.
}
