% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi2.R
\name{simulate_joint_pvalues}
\alias{simulate_joint_pvalues}
\title{Simulate p-values under the pisquared model with specified marginal pi0 and sharing rates.}
\usage{
simulate_joint_pvalues(N, pi_1, pi_2, true_jaccard, alpha, beta)
}
\arguments{
\item{N}{Sample size.}

\item{pi_1}{Marginal pi1 for "discovery" data.}

\item{pi_2}{Marginal pi1 for "replication" data.}

\item{true_jaccard}{Sharing level to aim for.}

\item{alpha}{First parameter of beta distribution on non-null p-values. Should be <1.}

\item{beta}{Second parameter of beta distribution on non-null p-values. Should be >=1.}
}
\value{
Data.frame with which elements are hits in each cohort and corresponding p values.
}
\description{
Simulate p-values under the pisquared model with specified marginal pi0 and sharing rates.
}
