% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi2.R
\name{simulate_pvalues}
\alias{simulate_pvalues}
\title{Simulate p-values for just one p-value array under the pisquared model.}
\usage{
simulate_pvalues(N, pi1, alpha, beta)
}
\arguments{
\item{N}{Sample size.}

\item{pi1}{Marginal pi1 (proportion of non-null p-values).}

\item{alpha}{First parameter of beta distribution on non-null p-values. Should be <1.}

\item{beta}{Second parameter of beta distribution on non-null p-values. Should be >=1.}
}
\value{
array of p-values
}
\description{
Simulate p-values for just one p-value array under the pisquared model.
}
