% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi2.R
\name{fit_truncated_beta}
\alias{fit_truncated_beta}
\title{Fit a beta distribution truncated to [0,upperlimit]}
\usage{
fit_truncated_beta(p, upper_limit, alpha_prior = c(2, 2),
  beta_prior = c(2, 1), method = "optimizing", ...)
}
\arguments{
\item{p}{Samples (p-values for pisquared's application)}

\item{upper_limit}{Upper limit of truncation (lower limit is fixed to 0)}

\item{alpha_prior}{Shape1 and shape2 of beta() prior on alpha. Shape1,shape2 > 1 prevents alpha collapsing to 0 or 1.}

\item{beta_prior}{Shape and rate of prior on beta_minus_one. Shape>1 enforces beta > 1.}

\item{method}{Can be "optimizing", "vb" or "sampling", determines approach for fitting stan model.}

\item{...}{passed to rstan fitting function choosen by `method`}
}
\value{
Fitted alpha and beta.
}
\description{
Fit a beta distribution truncated to [0,upperlimit]
}
